==> scripts/5_train_predictivemodel.py <==
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Created on Mon Jun 17 12:45:01 2019

@author: saracabellopinedo
"""

#import modules
from numpy.random import seed
seed(1)
from tensorflow import set_random_seed
set_random_seed(2)

import re
import pandas as pd
import numpy as np
from keras.models import load_model, Model
from keras.layers import Input, Dense

def load_subsets(original_path):
    #function to load train and test subsets for x and y:
    #x = environmental variables
    #y = OTU table (that in this models acts as the class to be predicted)
    
    regex = re.compile(".biom$")
    path = re.sub(regex, "", original_path)
    
    xtrain_file=path+"_xtrain80.csv"
    X_train = pd.read_csv(xtrain_file, sep='\t')
    ytrain_file=path+"_ytrain80.csv"
    y_train = pd.read_csv(ytrain_file, sep='\t')
    xtest_file=path+"_xtest20.csv"
    X_test = pd.read_csv(xtest_file, sep='\t')
    ytest_file=path+"_ytest20.csv"
    y_test = pd.read_csv(ytest_file, sep='\t')

    return X_train, y_train, X_test, y_test

def normalize_interval(data):
    #function for normalizing OTU table within [0, 1] interval
    #input: df with original values
    #output: df with normalized values
    
    #create empty df to allocate normalized data
    normalized_data=pd.DataFrame(columns=data.columns) 
    normalized_data['X.SampleID']=data['X.SampleID']
    normalized_data.index=normalized_data['X.SampleID']
    data.index=data['X.SampleID']
    
    #complete normalized_data df
    for row in data.itertuples(index=True, name='Pandas'):
        id_sample=row[1]
        values=[]
        values=np.array(data.loc[id_sample][1:len(row)+2])
        min_val=np.min(values)
        max_val=np.max(values)

        if (min_val != max_val):
            try:
                new_values=[(x - min_val)/ (max_val- min_val) for x in values]
            except:
                print("Strings found")
                new_values=values
        else:
            new_values=values
        

        new_values.insert(0, id_sample)
        normalized_data.at[id_sample]=new_values
    
    return normalized_data

def normalize_maximum(data):
    #normalize over the maximum value for each mapping variable
    
    normalized_data=pd.DataFrame(columns=data.columns)
    normalized_data['X.SampleID']=data['X.SampleID']
    for colname in data:
        if colname != 'X.SampleID':
            values=data[colname].tolist()
            maximum=max(values)
            if maximum > 1:
                try:
                    new_values=[x / maximum for x in values]
                except:
                    print("Se han encontrado strings")
                    new_values=values
            else:
                new_values=values
            
            normalized_data[colname]=new_values
    return normalized_data

def categorical2numbers(X_train, X_test):
    #transform each value of a categorical variable into a number
   
    n_train=len(X_train)
    metadata=X_train.append(X_test)
    
    #substitution by numbers
    if 'INBREDS' in metadata.columns:
        #dictionary that stores all inbreds possible values and its corresponding number
        dict_inbreds={'Oh7B':0, 'P39':1, 'CML333':2, 'Il14H':3, 'MS71':4, 'Oh43': 5, 'CML52':6, 
                      'Mo18W':7, 'M37W':8, 'Mo17':9, 'CML103':10, 'CML69':11, 'CML228':12,
                      'CML322':13, 'Tzi8':14, 'B73':15, 'NC350':16, 'CML247':17, 'B97':18,
                      'Ki3':19, 'Ky21':20, 'CML277':21, 'M162W':22, 'Hp301':23, 'NC358':24, 
                      'Tx303':25, 'Ki11':26}
        
        for key in dict_inbreds.keys():
            try:
                metadata=metadata.replace(key, dict_inbreds[key])
            except:
                pass
    
    if 'Maize_Line' in metadata.columns:
        #dictionary that stores maize line possible values and its corresponding number
        dict_maizeline={'Non_Stiff_Stalk':0, 'Sweet_Corn':1, 'Tropical':2, 'Mixed':3, 'Stiff_Stalk':4, 'Popcorn':5}   
        for key in dict_maizeline.keys():
            try:
                metadata=metadata.replace(key, dict_maizeline[key])
            except:
                pass
    
    X_train = metadata.iloc[0:n_train,:]
    X_test = metadata.iloc[n_train:len(metadata)+1,:]
    return X_train, X_test

def categorical2columns(X_train_df, X_test_df, n):
    #the n values found in a particular categorical variable are transformed into n columns 
    #which are filled using 0 (= not belonging to the variable level) or 1 (= belonging)
    
    n_train=len(X_train_df)
    metadata=X_train_df.append(X_test_df)

    #create columns for each value
    df_new=metadata.iloc[:, 0:n+1]
    metadata.index=metadata['X.SampleID']
    df_new.index=df_new['X.SampleID']
    
    if 'INBREDS' in metadata.columns:
        #column names created in the presence of variable Inbreds
        new_columns1 = ['INBREDS_Oh7B', 'INBREDS_P39', 'INBREDS_CML333', 'INBREDS_Il14H', 'INBREDS_MS71', 'INBREDS_Oh43', 'INBREDS_CML52', 
                      'INBREDS_Mo18W', 'INBREDS_M37W', 'INBREDS_Mo17', 'INBREDS_CML103', 'INBREDS_CML69', 'INBREDS_CML228',
                      'INBREDS_CML322', 'INBREDS_Tzi8', 'INBREDS_B73', 'INBREDS_NC350', 'INBREDS_CML247', 'INBREDS_B97',
                      'INBREDS_Ki3', 'INBREDS_Ky21', 'INBREDS_CML277', 'INBREDS_M162W', 'INBREDS_Hp301', 'INBREDS_NC358', 
                      'INBREDS_Tx303', 'INBREDS_Ki11']
        for column in new_columns1:
            df_new[column]=0
        
        for row in metadata.itertuples(index=True, name='Pandas'):
            id_sample=row[1]
            inbred=metadata.loc[id_sample, 'INBREDS']
            inbred_regexp=re.compile(str(".*" + str(inbred)))
            column_inbred=list(filter(inbred_regexp.match, new_columns1))[0]
            df_new.at[id_sample, column_inbred] = 1
            
    
    if 'Maize_Line' in metadata.columns:
        #column names created in the presence of variable Maize_Line
        new_columns2 = ['LINE_Non_Stiff_Stalk', 'LINE_Sweet_Corn', 'LINE_Tropical', 'LINE_Mixed', 
                      'LINE_Stiff_Stalk', 'LINE_Popcorn']
        for column in new_columns2:
            df_new[column]=0
         
        for row in metadata.itertuples(index=True, name='Pandas'):
            id_sample=row[1]
            maize_line=metadata.loc[id_sample, 'Maize_Line']
            maizeline_regexp=re.compile(str(".*" + str(maize_line)))
            column_maizeline=list(filter(maizeline_regexp.match, new_columns2))[0]
            df_new.at[id_sample, column_maizeline] = 1
            
    metadata=df_new   
    X_train = metadata.iloc[0:n_train,:]
    X_test = metadata.iloc[n_train:len(metadata)+1,:]
    return X_train, X_test

def categorical2binary(X_train, X_test, n):
    #each value of a particular categorical variable is codified using binary codification
    #and n new columns are created were n = number of necessary binary digits
    n_train=len(X_train)
    metadata=X_train.append(X_test)

    #create columns for each value
    df_new=metadata.iloc[:, 0:n+1]
    metadata.index=metadata['X.SampleID']
    df_new.index=df_new['X.SampleID']
    
    if 'INBREDS' in metadata.columns:
        #binary codification for values of inbreds
        dict_inbreds={'Oh7B':'00001', 'P39':'00010', 'CML333':'00011', 'Il14H':'00100', 'MS71':'00101', 'Oh43': '00110', 'CML52':'00111', 
                      'Mo18W':'01000', 'M37W':'01001', 'Mo17':'01010', 'CML103':'01011', 'CML69':'01100', 'CML228':'01101',
                      'CML322':'01110', 'Tzi8':'01111', 'B73':'10000', 'NC350':'10001', 'CML247':'10010', 'B97':'10011',
                      'Ki3':'10100', 'Ky21':'10101', 'CML277':'10110', 'M162W':'10111', 'Hp301':'11000', 'NC358':'11001', 
                      'Tx303':'11010', 'Ki11':'11011'}
        #5 digits required = 5 new columns created, one containing each digit 
        new_columns1 = ['INBREDS_4', 'INBREDS_3', 'INBREDS_2', 'INBREDS_1', 'INBREDS_0']
        
        for column in new_columns1:
            df_new[column]=0

        for row in metadata.itertuples(index=True, name='Pandas'):
            id_sample=row[1]
            inbred=metadata.loc[id_sample, 'INBREDS']
            code_inbred=list(dict_inbreds[inbred])
            df_new.at[id_sample, 'INBREDS_4'] = code_inbred[0]
            df_new.at[id_sample, 'INBREDS_3'] = code_inbred[1]
            df_new.at[id_sample, 'INBREDS_2'] = code_inbred[2]
            df_new.at[id_sample, 'INBREDS_1'] = code_inbred[3]
            df_new.at[id_sample, 'INBREDS_0'] = code_inbred[4]   
            
    if 'Maize_Line' in metadata.columns:
        #codification for the different maize lines found in data
        dict_maizeline={'Non_Stiff_Stalk':'001', 'Sweet_Corn':'010', 'Tropical':'011', 'Mixed':'100', 'Stiff_Stalk':'101', 'Popcorn':'110'}   
        
        #3 new columns
        new_columns2 =[ 'LINE_2', 'LINE_1', 'LINE_0']
        
        for column in new_columns2:
            df_new[column]=0
    
        for row in metadata.itertuples(index=True, name='Pandas'):
            id_sample=row[1]
            maize_line=metadata.loc[id_sample, 'Maize_Line']
            code_maizeline=list(dict_maizeline[maize_line])
            df_new.at[id_sample, 'LINE_2'] = code_maizeline[0]    
            df_new.at[id_sample, 'LINE_1'] = code_maizeline[1]    
            df_new.at[id_sample, 'LINE_0'] = code_maizeline[2]    
    
    metadata=df_new   
    
    X_train = metadata.iloc[0:n_train,:]
    X_test = metadata.iloc[n_train:len(metadata)+1,:]
    return X_train, X_test

def transform_data(X_train_df, X_test_df, y_train_df, y_test_df, option, n):
    #function to transform original data. 2 types of transformation are performed here:
    # 1. categorical variables tranformation
    # 2. normalization of data to train the model: OTU table (to interval [0, 1]) and mapping variables (dividing into the maximum value)
    #output: 4 df
    
    if option==1:
        #substitution by numbers --> option 1
        X_train_df, X_test_df=categorical2numbers(X_train_df, X_test_df)
    elif option==2:
        #create colums for each value--> option 2
        X_train_df, X_test_df=categorical2columns(X_train_df, X_test_df, n-1)
    elif option==3:
        #create columns with binary coding --> option 3
        X_train_df, X_test_df=categorical2binary(X_train_df, X_test_df, n-1)
    
    X_test_df = normalize_maximum(X_test_df)
    X_train_df = normalize_maximum(X_train_df)
    y_test_df = normalize_interval(y_test_df)
    y_train_df = normalize_interval(y_train_df)
    
    return X_train_df, X_test_df, y_train_df, y_test_df

def format_data(X_train, y_train, X_test, y_test):
    #function to transform the 4 df required to train the model to fit the requirements of
    #training cycles of deep learning in keras
    
    #removing columns that contain the id of the sample, not interesting for the model
    if 'X.SampleID' in X_train:
        X_train.drop(['X.SampleID'], axis='columns', inplace=True)

    if 'X.SampleID' in X_test:
        X_test.drop(['X.SampleID'], axis='columns', inplace=True) 
        
    if 'X.SampleID' in y_train:
        y_train.drop(['X.SampleID'], axis='columns', inplace=True)
        
    if 'X.SampleID' in y_test:
        y_test.drop(['X.SampleID'], axis='columns', inplace=True) 
        
    #transforming df into list of lists (format requiered by keras models) 
    if not isinstance(X_train, np.ndarray):
        X_train_list=[]
        for row in X_train.itertuples(index=True, name='Pandas'):
                X_train_list.append(row[1:])
        X_train = np.asarray(X_train)
    
    if not isinstance(X_test, np.ndarray):
        X_test_list=[]
        for row in X_test.itertuples(index=True, name='Pandas'):
                X_test_list.append(row[1:])  
        X_test = np.asarray(X_test_list)
        
    if not isinstance(y_train, np.ndarray):
        y_train_list=[]
        for row in y_train.itertuples(index=True, name='Pandas'):
            y_train_list.append(row[1:])
        y_train=np.asarray(y_train_list)
    
    if not isinstance(y_test, np.ndarray):
        y_test_list=[]
        for row in y_test.itertuples(index=True, name='Pandas'):
            y_test_list.append(row[1:])  
        y_test = np.asarray(y_test_list)
    
    return X_train, y_train, X_test, y_test

def load_autoencoder(path):
    #load autoencoder model and it parts: encoder + decoder. 
    #This particular functions is useful only for architecture 4 and 5
    
    autoencoder = load_model(path)
    encoder = Model(autoencoder.input, autoencoder.layers[-4].output)
    encoding_dim=encoder.layers[-1].output_shape[1]
    decoder_input = Input(shape=(encoding_dim,))
    deco = autoencoder.layers[-3](decoder_input)
    deco = autoencoder.layers[-2](deco)
    deco = autoencoder.layers[-1](deco)
    decoder = Model(decoder_input, deco)
    
    return autoencoder, encoder, decoder

#train predictive models with variable number of hidden layers (indicated in the name of the function)
#between the input layer and the coding layer
#They return the model itself and also information about error measurements obtained during training cycles
    
def train_model_0hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder):
    y_train_encoded = encoder.predict(y_train)
    y_test_encoded = encoder.predict(y_test)
    
    # model definition
    original_dim=len(X_train[0])
    encoding_dim = len (y_train_encoded[0])
    
    input_layer= Input(shape=(original_dim,))
    
    encoded = Dense (encoding_dim, activation=activation_function)(input_layer)
    model = Model (input_layer, encoded)
    model.compile(optimizer='adam', loss='mse')
    
    seed(1)
    set_random_seed(2)
    result=model.fit(X_train, y_train_encoded,
                    nb_epoch=n_iter,
                    batch_size=256,
                    shuffle=True,
                    validation_data=(X_test, y_test_encoded),
                    verbose=1)

    return model, result.history   

def train_model_1hiddenlayer(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, hidden_dim):
    y_train_encoded = encoder.predict(y_train)
    y_test_encoded = encoder.predict(y_test)
    
    #   model definition
    original_dim=len(X_train[0])
    encoding_dim = len (y_train_encoded[0])
    
    input_layer= Input(shape=(original_dim,))
    
    encoded1 = Dense (hidden_dim, activation=activation_function)(input_layer)
    encoded2 = Dense (encoding_dim, activation=activation_function)(encoded1)
    model = Model (input_layer, encoded2)
    model.compile(optimizer='adam', loss='mse')
    
    seed(1)
    set_random_seed(2)
    result=model.fit(X_train, y_train_encoded,
                    nb_epoch=n_iter,
                    batch_size=256,
                    shuffle=True,
                    validation_data=(X_test, y_test_encoded),
                    verbose=1)

    return model, result.history   

def train_model_2hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, hidden_dim1, hidden_dim2):
    y_train_encoded = encoder.predict(y_train)
    y_test_encoded = encoder.predict(y_test)
    
    #   model definition
    original_dim=len(X_train[0])
    encoding_dim = len (y_train_encoded[0])
    
    input_layer= Input(shape=(original_dim,))
    
    encoded1 = Dense (hidden_dim1, activation=activation_function)(input_layer)
    encoded2 = Dense (hidden_dim2, activation=activation_function)(encoded1)
    encoded3 = Dense (encoding_dim, activation=activation_function)(encoded2)
    model = Model (input_layer, encoded3)
    model.compile(optimizer='adam', loss='mse')
    
    seed(1)
    set_random_seed(2)
    result=model.fit(X_train, y_train_encoded,
                    nb_epoch=n_iter,
                    batch_size=256,
                    shuffle=True,
                    validation_data=(X_test, y_test_encoded),
                    verbose=1)

    return model, result.history   

def train_model_3hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, hidden_dim1, hidden_dim2, hidden_dim3):
    y_train_encoded = encoder.predict(y_train)
    y_test_encoded = encoder.predict(y_test)
    
    #   model definition
    original_dim=len(X_train[0])
    encoding_dim = len (y_train_encoded[0])
    
    input_layer= Input(shape=(original_dim,))
    
    encoded1 = Dense (hidden_dim1, activation=activation_function)(input_layer)
    encoded2 = Dense (hidden_dim2, activation=activation_function)(encoded1)
    encoded3 = Dense (hidden_dim3, activation=activation_function)(encoded2)
    encoded4 = Dense (encoding_dim, activation=activation_function)(encoded3)
    model = Model (input_layer, encoded4)
    model.compile(optimizer='adam', loss='mse')
    
    seed(1)
    set_random_seed(2)
    result=model.fit(X_train, y_train_encoded,
                    nb_epoch=n_iter,
                    batch_size=256,
                    shuffle=True,
                    validation_data=(X_test, y_test_encoded),
                    verbose=1)

    return model, result.history   

def save_results(X_train_df, y_train_df, X_test_df, y_test_df, n_iter, df_results, encoder, decoder, architecture):
    #save error measurements of each training assay and the number of iterations in which the minimum value is found
    
    #obtain data in the correct format for keras models input
    X_train, y_train, X_test, y_test = format_data(X_train_df, y_train_df, X_test_df, y_test_df)

    activation_function='tanh'
    
    model, result = train_model_0hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder)
    
    min_value_valloss = min(result['val_loss']) #minimum error within test subset
    iter_min = result['val_loss'].index(min_value_valloss) + 1 #when this minimum is found
    min_value_loss = result['loss'][iter_min-1] #mse value in train subset at that time
    
    ind = len(df_results.index) + 1
    splitted = architecture.split('-')
    splitted = splitted[-3:len(splitted)]
    s = "-"
    model_name = s.join(splitted)
    #save results in a dataframe
    df_results.loc[ind] = [architecture, model_name, min_value_loss, min_value_valloss, iter_min]
    df_results.to_csv('./results_prediction.csv', sep='\t')
    
    #different values that size of hidden layers can take
    dims=[4, 5, 6, 10, 20]
    #models with only 1 hidden layer between input and code
    for dim in dims:
        model, result = train_model_1hiddenlayer(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, dim)
        min_value_valloss = min(result['val_loss'])
        iter_min = result['val_loss'].index(min_value_valloss) + 1
        min_value_loss = result['loss'][iter_min-1]
        
        ind = len(df_results.index) + 1
        submodel_name = str(dim) + "-" + model_name
        df_results.loc[ind] = [architecture, submodel_name, min_value_loss, min_value_valloss, iter_min]
        df_results.to_csv('./results_prediction.csv', sep='\t')
    
    #models with two hidden layers between input and code
    for dim1 in dims:
        for dim2 in dims:
            model, result = train_model_2hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, dim1, dim2)
            min_value_valloss = min(result['val_loss'])
            iter_min = result['val_loss'].index(min_value_valloss) + 1
            min_value_loss = result['loss'][iter_min-1]
             
            ind = len(df_results.index) + 1
            submodel_name = str(dim1) + "-" + str(dim2) + "-" + model_name
            df_results.loc[ind] = [architecture, submodel_name, min_value_loss, min_value_valloss, iter_min]
            df_results.to_csv('./results_prediction.csv', sep='\t')
            
    #models with 3 hidden layer between input and code
    for dim1 in dims:
        for dim2 in dims:
            for dim3 in dims:
                model, result = train_model_3hiddenlayers(X_train, y_train, X_test, y_test, n_iter, activation_function, encoder, dim1, dim2, dim3)
                min_value_valloss = min(result['val_loss'])
                iter_min = result['val_loss'].index(min_value_valloss) + 1
                min_value_loss = result['loss'][iter_min-1]
                 
                ind = len(df_results.index) + 1
                submodel_name = str(dim1) + "-" + str(dim2) + "-" +  str(dim3) + "-" + model_name
                df_results.loc[ind] = [architecture, submodel_name, min_value_loss, min_value_valloss, iter_min]
                df_results.to_csv('./results_prediction.csv', sep='\t')
    
    return (df_results)   

#create empty df to store training results
df_results=pd.DataFrame(columns=['autoencoder', 'model', 'mse_train', 'mse_test', 'n_iter'])

#load data to train the model
X_train_df, y_train_df, X_test_df, y_test_df = load_subsets("./data/original_data/2otu_table2save_80.biom")
n=len(X_train_df.columns) - 2

#transform original data
X_train_df, X_test_df, y_train_df, y_test_df = transform_data(X_train_df, X_test_df, y_train_df, y_test_df, 1, n)

#testing the models
architectures2test = ['100-10-6-100-500', '100-10-6-50-500', '50-10-6-50-500', '20-10-6-100-1000', '10-6-100-500', '10-6-100-1000']
arch= ['architecture4', 'architecture4', 'architecture4', 'architecture4', 'architecture5', 'architecture5']
number= ['4', '4', '4', '4', '5', '5'] #variables to build the path to find the autoencoder model

for i in range(len(architectures2test)):
    path ="./models/predictive_models/" + str(arch[i]) + '/autoencoder2_' + str(number[i]) + '_' + str(architectures2test[i]) + '.hdf5'
    
    #load autoencoder model and its encoder and decoder parts. We are going to use the decoder part to build the predictive model
    autoencoder, encoder, decoder = load_autoencoder(path)
    df_results = save_results(X_train_df, y_train_df, X_test_df, y_test_df, 30000 , df_results, encoder, decoder, architectures2test[i])
